requirements.txt
python-pptx>=0.6.21
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
httpx==0.27.0
graphviz==0.20.1
matplotlib==3.8.2